Revert "Make room for adding more colorspaces to ffdshow"
This reverts commit 6eff38c8d15e76efe5f6bb57159bb688a7e0ec60.

+++ src/TffdshowEnc.h
@@ -125,7 +125,7 @@ protected:
     unsigned int dx,dy;
     unsigned int outDx,outDy;
     Tcsps enccsps;
-    uint64_t enccsp;
+    int enccsp;
     Tconvert *convert;
     int keyspacing;
     int getQuantQuant(void);
@@ -137,7 +137,7 @@ protected:
     bool firstrun;
 
     HWND h_graph;
-    uint64_t inColorspace;
+    int inColorspace;
     BITMAPINFO biOutput;
     TencStats encStats;
