+++ src/imgFilters/TimgFilterSubtitles.h
@@ -125,7 +125,7 @@ HANDLE getGlyphThreadHandle();
 
 protected:
 virtual bool is(const TffPictBase &pict,const TfilterSettingsVideo *cfg);
-virtual uint64_t getSupportedInputColorspaces(const TfilterSettingsVideo *cfg) const
+virtual int getSupportedInputColorspaces(const TfilterSettingsVideo *cfg) const
 {
     return FF_CSP_420P;
 }
